Fix crash that would occur when actual number of sections differed to expected during a batch update. Reload the collection-view if there is a mismatch.
Tested this with two participants firing random messages every ~1 second. With this change the crash did not occur and the reloadData was hit about as often as the crash was occurring before the change.

This test also demonstrated a deadlock in the LayerKit code. It appears that even read-only accesses require locks.